Extract language query runner partial
